@@ -1,11 +1,14 @@
 module github.com/digitalocean/gocop
 
-go 1.12
+go 1.19
 
 require (
-	github.com/inconshreveable/mousetrap v1.0.0 // indirect
-	github.com/lib/pq v1.2.0
-	github.com/poy/onpar v0.0.0-20181125144932-f2f06780798d
-	github.com/spf13/cobra v0.0.3
-	github.com/spf13/pflag v1.0.3 // indirect
+	github.com/lib/pq v1.10.7
+	github.com/poy/onpar v1.1.2
+	github.com/spf13/cobra v1.6.1
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 )
